+++ src/renderer/components/swap/Swap.tsx
@@ -277,6 +277,7 @@ export const Swap = ({
   const intl = useIntl()
 
   const { chain: sourceChain } = sourceAsset.type === AssetType.SYNTH ? dex.asset : sourceAsset
+  const { chain: targetChain } = targetAsset.type === AssetType.SYNTH ? dex.asset : targetAsset
 
   const lockedWallet: boolean = useMemo(() => isLocked(keystore) || !hasImportedKeystore(keystore), [keystore])
   const [quoteOnly, setQuoteOnly] = useState<boolean>(false)
@@ -318,8 +319,8 @@ export const Swap = ({
   const [enabledChains, setEnabledChains] = useState<Set<EnabledChain>>(new Set())
   const [disabledChains, setDisabledChains] = useState<EnabledChain[]>([])
 
-  const isTargetChainDisabled = disabledChains.includes(targetAsset.chain)
-  const isSourceChainDisabled = disabledChains.includes(sourceAsset.chain)
+  const isTargetChainDisabled = disabledChains.includes(targetChain)
+  const isSourceChainDisabled = disabledChains.includes(sourceChain)
 
   useEffect(() => {
     const subscription = userChains$.subscribe((chains: EnabledChain[]) => {