+++ src/renderer/components/wallet/txs/send/SendFormEVM.tsx
@@ -22,11 +22,12 @@ import { Form } from 'antd'
 import Tooltip from 'antd/es/tooltip'
 import { RadioChangeEvent } from 'antd/lib/radio'
 import BigNumber from 'bignumber.js'
+import * as A from 'fp-ts/lib/Array'
 import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
 import { useIntl } from 'react-intl'
 
-import { Dex } from '../../../../../shared/api/types'
+import { Dex, TrustedAddress, TrustedAddresses } from '../../../../../shared/api/types'
 import { chainToString, isChainOfMaya } from '../../../../../shared/utils/chain'
 import { isKeystoreWallet, isLedgerWallet } from '../../../../../shared/utils/guard'
 import { WalletType } from '../../../../../shared/wallet/types'
@@ -80,6 +81,7 @@ export type FormValues = {
 
 export type Props = {
   asset: SelectedWalletAsset
+  trustedAddresses: TrustedAddresses | undefined
   balances: WalletBalances
   balance: WalletBalance
   transfer$: SendTxStateHandler
@@ -99,6 +101,7 @@ export type Props = {
 export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
   const {
     asset: { walletType, walletAccount, walletIndex, hdMode, walletAddress },
+    trustedAddresses,
     poolDetails,
     balances,
     balance,
@@ -163,6 +166,53 @@ export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
   const [currentMemo, setCurrentMemo] = useState<string>('')
   const [affiliateTracking, setAffiliateTracking] = useState<string>('')
   const isChainAsset = isEthAsset(asset) || isAvaxAsset(asset) || isBscAsset(asset) || isAethAsset(asset)
+
+  const oSavedAddresses: O.Option<TrustedAddress[]> = useMemo(
+    () =>
+      FP.pipe(O.fromNullable(trustedAddresses?.addresses), O.map(A.filter((address) => address.chain === asset.chain))),
+    [trustedAddresses, asset.chain]
+  )
+
+  const handleSavedAddressSelect = useCallback(
+    (value: string) => {
+      form.setFieldsValue({ recipient: value })
+      setRecipientAddress(value)
+      if (value) {
+        const matched = FP.pipe(
+          oSavedAddresses,
+          O.map((addresses) => addresses.filter((address) => address.address.includes(value))),
+          O.chain(O.fromPredicate((filteredAddresses) => filteredAddresses.length > 0))
+        )
+        setMatchedAddresses(matched)
+      }
+    },
+    [form, oSavedAddresses]
+  )
+
+  const renderSavedAddressesDropdown = useMemo(
+    () =>
+      FP.pipe(
+        oSavedAddresses,
+        O.fold(
+          () => null,
+          (addresses) => (
+            <Form.Item label={intl.formatMessage({ id: 'common.savedAddresses' })} className="mb-20px">
+              <Styled.CustomSelect
+                placeholder={intl.formatMessage({ id: 'common.savedAddresses' })}
+                onChange={(value) => handleSavedAddressSelect(value as string)}
+                style={{ width: '100%' }}>
+                {addresses.map((address) => (
+                  <Styled.CustomSelect.Option key={address.address} value={address.address}>
+                    {address.name}: {address.address}
+                  </Styled.CustomSelect.Option>
+                ))}
+              </Styled.CustomSelect>
+            </Form.Item>
+          )
+        )
+      ),
+    [oSavedAddresses, intl, handleSavedAddressSelect]
+  )
 
   const handleMemo = useCallback(() => {
     let memoValue = form.getFieldValue('memo') as string
@@ -565,10 +615,18 @@ export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
         return routerAddress === address
       } // If Some, compare the address
     )(routerOption)
-
+  const [matchedAddresses, setMatchedAddresses] = useState<O.Option<TrustedAddress[]>>(O.none)
   const onChangeAddress = useCallback(
     async ({ target }: React.ChangeEvent<HTMLInputElement>) => {
       const address = target.value
+      if (address) {
+        const matched = FP.pipe(
+          oSavedAddresses,
+          O.map((addresses) => addresses.filter((address) => address.address.includes(address.address))),
+          O.chain(O.fromPredicate((filteredAddresses) => filteredAddresses.length > 0)) // Use O.none for empty arrays
+        )
+        setMatchedAddresses(matched)
+      }
       // we have to validate input before storing into the state
       const isNotAllowed = checkAddress(routers.MAYA, address) || (checkAddress(routers.THOR, address) && !isChainAsset)
       setNotAllowed(isNotAllowed)
@@ -579,7 +637,7 @@ export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
         })
         .catch(() => setSendAddress(O.none))
     },
-    [addressValidator, isChainAsset, routers.MAYA, routers.THOR]
+    [addressValidator, isChainAsset, oSavedAddresses, routers.MAYA, routers.THOR]
   )
 
   const reloadFees = useCallback(() => {
@@ -873,7 +931,10 @@ export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
     [balances, recipientAddress]
   )
 
-  const renderWalletType = useMemo(() => H.renderedWalletType(oMatchedWalletType), [oMatchedWalletType])
+  const renderWalletType = useMemo(
+    () => H.renderedWalletType(oMatchedWalletType, matchedAddresses),
+    [matchedAddresses, oMatchedWalletType]
+  )
 
   return (
     <>
@@ -890,6 +951,7 @@ export const SendFormEVM: React.FC<Props> = (props): JSX.Element => {
           onFinish={() => setShowConfirmationModal(true)}
           labelCol={{ span: 24 }}>
           <Styled.SubForm>
+            {renderSavedAddressesDropdown}
             <Styled.CustomLabel size="big">
               {intl.formatMessage({ id: 'common.address' })}
               {renderWalletType}