+++ src/renderer/i18n/types.ts
@@ -155,6 +155,8 @@ export type CommonMessageKey =
   | 'common.tx.type.donate'
   | 'common.tx.type.deposit'
   | 'common.tx.type.withdraw'
+  | 'common.tx.type.send'
+  | 'common.tx.type.runePoolDeposit'
   | 'common.time.days'
   | 'common.time.days.short'
   | 'common.time.month1'