+++ src/renderer/services/midgard/action.utils.ts
@@ -15,7 +15,8 @@ export const getTxType = (apiString: string): TxType => {
     case 'WITHDRAW':
     case 'DONATE':
     case 'REFUND':
-    case 'SWITCH':
+    case 'SEND':
+    case 'RUNEPOOLDEPOSIT':
       return type
     case 'ADDLIQUIDITY':
       return 'DEPOSIT'
@@ -31,7 +32,7 @@ export const getRequestType = (type?: TxType | 'ALL'): string | undefined => {
     case 'DEPOSIT': {
       return 'addLiquidity'
     }
-    case 'SWITCH':
+    case 'SEND':
     case 'SWAP':
     case 'WITHDRAW':
     case 'DONATE':